+++ app/src/main/java/team/retum/jobisandroidv2/root/RootScreen.kt
@@ -24,15 +24,15 @@ import team.retum.bookmark.navigation.bookmarks
 import team.retum.home.R
 import team.retum.home.navigation.NAVIGATION_HOME
 import team.retum.home.navigation.home
+import team.retum.jobis.recruitment.navigation.navigateToRecruitmentDetails
+import team.retum.jobis.recruitment.navigation.navigateToRecruitments
+import team.retum.jobis.recruitment.navigation.recruitments
 import team.retum.jobisandroidv2.ui.BottomNavigationBar
 import team.retum.jobisdesignsystemv2.button.ButtonColor
 import team.retum.jobisdesignsystemv2.button.JobisButton
 import team.retum.jobisdesignsystemv2.foundation.JobisTheme
 import team.retum.jobisdesignsystemv2.foundation.JobisTypography
 import team.retum.jobisdesignsystemv2.text.JobisText
-import team.retum.recruitment.navigation.navigateToRecruitmentDetails
-import team.retum.recruitment.navigation.navigateToRecruitments
-import team.retum.recruitment.navigation.recruitments
 import team.returm.mypage.navigation.mypage
 
 @OptIn(ExperimentalMaterialApi::class)
