+++ feature/recruitment/src/main/java/team/retum/jobis/recruitment/ui/RecruitmentDetailsScreen.kt
@@ -1,4 +1,4 @@
-package team.retum.recruitment.ui
+package team.retum.jobis.recruitment.ui
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.Image
@@ -32,6 +32,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import team.retum.common.enums.HiringProgress
+import team.retum.jobis.recruitment.R
 import team.retum.jobisdesignsystemv2.appbar.JobisSmallTopAppBar
 import team.retum.jobisdesignsystemv2.button.ButtonColor
 import team.retum.jobisdesignsystemv2.button.JobisButton
@@ -41,7 +42,6 @@ import team.retum.jobisdesignsystemv2.foundation.JobisTheme
 import team.retum.jobisdesignsystemv2.foundation.JobisTypography
 import team.retum.jobisdesignsystemv2.text.JobisText
 import team.retum.jobisdesignsystemv2.utils.clickable
-import team.retum.recruitment.R
 
 // TODO: 서버 연동 시 제거
 private data class RecruitmentDetail(
